backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
channels>=4.1
channels-redis>=4.2
psycopg[binary]>=3.2
python-dotenv>=1.0
